backend/requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.0
Django==5.1
django-ckeditor-5==0.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
packaging==24.2
pillow==10.4.0
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==21.10
sniffio==1.3.1
sqlparse==0.5.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
